requirements.txt
requests
websocket-client
backports.ssl_match_hostname
#certifi
